Updates the testing note section in the template
Test see the testing notes as confirmation from their point of view on what is likely to need testing. Currently dev's see this section in the issue and detail notes from their point of view, this means the "testing notes" have different meanings to different disciplines. Therefore I'm clarifying that the notes shouldn't be confirmed and completed by testers and dev can give insight into how the issue is likely to affect the application, which should help the tester in defining the testing notes.