Add tick box to prompt people to assign PR/issue
When using the github projects board it can be confusing who is working on an issue/PR as the creator of the item is displayed. This can become misleading when person X created an issue and person Y picked up the issue. Therefore adding a tick box to prompt people to assign themselves.