Remove double run as it's obsolete
Revisiting this and I've realised that the race condition will never be a factor. This is because I was comparing the trigger of a job against the checking of the API, whereas the instant check of the API is irrelevant given the fact github enforces we are up-to-date with latest.

Example:
PR 1 is merged at the same time as PR 2 is pushed. PR 2 triggers `can-i-merge` and is in the small timeframe that it will get a false-positive tick. Several seconds later Github enforces that the PR needs to pull in latest before merge and in doing so the `can-i-merge` job would be retriggered. Problem resolved.