Rename CPSPageGrid to CPSGrid
The name CPSPageGrid suggests that it will be used on a page level.
This might not be true in some use cases. i.e We might want to use the
Grid on a FrontPageSection which is not a page level component.

This might be confusing because it is named a `PageGrid` but not being
used at the page level. Naming it CPSGrid clears this confusion