Use new Cookie Oven URL
We have a replacement for the cookie oven and would like to turn off the old one.

This PR also includes hitting the oven for both first and third party so Safari does not delete the JS-set cookie after 7 days.

Also, we stop hitting the oven when setting the default opt-out as an optimisation since we normally only hit the oven as part of an explicit choice. In fact, the new endpoint sets explicit for you.