FIX: pass tests by removing test
Proves the caching works, I guess. When we tried to run tests for data fetch fail, it returned data because it was warm in the cache. If we think this is a good test to run, we can either pass a var into the fetch data function that will do _something_ or we can maybe run the test earlier so it doesn't have a warm cache