INI: terrible code to test caching generic service responses (#10606)
* INI: terrible code to test caching generic service responses

* FIX: remove console log tests

* FIX: linting

* FIX: is this an antipattern?

* FIX: test suites seemed to lose process.env vars

Attempt to get tests to pass then try and figure out why tests can't seem to access the env vars

* FIX: pass tests by removing test

Proves the caching works, I guess. When we tried to run tests for data fetch fail, it returned data because it was warm in the cache. If we think this is a good test to run, we can either pass a var into the fetch data function that will do _something_ or we can maybe run the test earlier so it doesn't have a warm cache

* Update src/app/routes/cpsAsset/utils/getAdditionalPageData.js

Co-authored-by: Karina Thomas <58214768+karinathomasbbc@users.noreply.github.com>

* FIX: that didn't work

* Cache all secondary data requests except recommendations

* Oops (remove console log)

---------

Co-authored-by: Karina Thomas <58214768+karinathomasbbc@users.noreply.github.com>
Co-authored-by: karinathomasbbc <karina.thomas@bbc.co.uk>